Convert WPP to NET Core 3.1 (#530)
* Convert WPP to NET Core 3.1

- Copyright Update
- Hotfix Parsing
- DB2Hash update
- Update WDC3 Reader
Thanks goes to @wowdev for DBC library

* Obviously point to correct module path

* Fix SQL Database Reading

* Fix Broadcast text loading

* Rename ExecuteQuery to CreateCommand

* Travis Update

Re-Add travis-build.sh
Create travis-test.sh

* Fix BitTest for NUnit

* Fix Drag&Drop from external directory

* Fix DateTime issues

* Remove useless line in test

* Remove macOS from appveyor.yml

.NET Core 3.1 not supported yet at appveyor for macOS

* Fix output path

* Add back macOS as appveyor has .NET Core 3.1 SDK installed

+++ WowPacketParser/DBC/DBC.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.IO;
 using System.Reflection;
 using System.Threading.Tasks;
-using DB2FileReaderLib.NET;
+using DBFileReaderLib;
 using WowPacketParser.DBC.Structures.BattleForAzeroth;
 using WowPacketParser.Misc;
 
@@ -114,7 +114,7 @@ namespace WowPacketParser.DBC
                             MapSpawnMaskStores.Add(mapDifficulty.Value.MapID, difficultyID);
 
                         if (!MapDifficultyStores.ContainsKey(mapDifficulty.Value.MapID))
-                            MapDifficultyStores.Add(mapDifficulty.Value.MapID, new List<byte>() { mapDifficulty.Value.DifficultyID });
+                            MapDifficultyStores.Add(mapDifficulty.Value.MapID, new List<int>() { mapDifficulty.Value.DifficultyID });
                         else
                             MapDifficultyStores[mapDifficulty.Value.MapID].Add(mapDifficulty.Value.DifficultyID);
                     }
@@ -208,7 +208,7 @@ namespace WowPacketParser.DBC
 
         public static readonly Dictionary<uint, string> Zones = new Dictionary<uint, string>();
         public static readonly Dictionary<int, int> MapSpawnMaskStores = new Dictionary<int, int>();
-        public static readonly Dictionary<int, List<byte>> MapDifficultyStores = new Dictionary<int, List<byte>>();
+        public static readonly Dictionary<int, List<int>> MapDifficultyStores = new Dictionary<int, List<int>>();
         public static readonly Dictionary<ushort, string> CriteriaStores = new Dictionary<ushort, string>();
         public static readonly Dictionary<uint, FactionEntry> FactionStores = new Dictionary<uint, FactionEntry>();
         public static readonly Dictionary<Tuple<uint, uint>, SpellEffectEntry> SpellEffectStores = new Dictionary<Tuple<uint, uint>, SpellEffectEntry>();
